Rename ReportPage command to show community page, add test for new method of ShowInBrowserService
